// types/DroneData.tsx
export interface DroneData {
    id?: number;
    name: string;
    weight: number;
    distance: number;
    status: string;
    status_battery: string;
    status_fly: string;
    status_order: string;
    latitude: string;
    longitude: string;
    battery: number;
    created_at: string;
    updated_at: string;
}